+++ Cargo.toml
@@ -20,7 +20,8 @@ version.workspace = true
 workspace = true
 
 [features]
-default = ["client", "server"]
+default = ["client"]
+# default = ["client", "server"]
 # Includes DWN client features such as request builders.
 client = []
 # Includes DWN server features such handlers.
@@ -34,6 +35,8 @@ base64ct.workspace = true
 chrono = { version = "0.4.39", features = ["serde"] }
 ciborium = "0.2.2"
 cid.workspace = true
+credibil-did.workspace = true
+credibil-infosec.workspace = true
 derive_more = { version = "1.0.0", features = ["display"] }
 ed25519-dalek.workspace = true
 futures.workspace = true
@@ -51,8 +54,6 @@ thiserror = "2.0.11"
 tokio.workspace = true
 ulid = "1.1.4"
 uuid = { version = "1.13.1", features = ["fast-rng", "v4"] }
-credibil-did.workspace = true
-credibil-infosec.workspace = true
 x25519-dalek.workspace = true
 zeroize = "1.8.1"
 
@@ -93,6 +94,8 @@ nursery = "warn"
 anyhow = "1.0.95"
 base64ct = { version = "1.6.0", features = ["std"] }
 cid = { version = "0.11.1", features = ["alloc"] }
+credibil-did = { git = "https://github.com/credibil/did.git" }
+credibil-infosec = { git = "https://github.com/credibil/infosec.git" }
 dwn-node = { path = "." }
 ed25519-dalek = { version = "2.1.1", features = ["rand_core"] }
 futures = "0.3.31"
@@ -105,8 +108,6 @@ serde_json = { version = "1.0.138", features = ["alloc"] }
 sha2 = "0.10.8"
 test-node = { path = "crates/test-node" }
 tokio = { version = "1.43.0", features = ["macros", "rt-multi-thread", "time"] }
-credibil-did = { git = "https://github.com/credibil/did.git" }
-credibil-infosec = { git = "https://github.com/credibil/infosec.git" }
 x25519-dalek = { version = "2.0.1", features = ["static_secrets"] }
 
 # [patch."https://github.com/credibil/infosec.git"]
