+++ src/grants/verify.rs
@@ -8,8 +8,8 @@ use chrono::{DateTime, Utc};
 use super::{RecordsScope, Scope};
 use crate::grants::{Grant, Publication};
 use crate::interfaces::Descriptor;
+use crate::interfaces::protocols::REVOCATION_PATH;
 use crate::interfaces::records::{Delete, Query, Read, RecordsFilter, Subscribe, Write};
-use crate::protocols::REVOCATION_PATH;
 use crate::provider::MessageStore;
 use crate::store::RecordsQueryBuilder;
 use crate::{Result, forbidden};
