+++ src/handlers.rs
@@ -1,0 +1,16 @@
+//! # Handlers
+//!
+//! This module contains the DWN message handlers — one for each
+//! interface/method message.
+
+mod authorize;
+mod messages_query;
+mod messages_read;
+mod messages_subscribe;
+mod protocols_configure;
+mod protocols_query;
+mod records_delete;
+mod records_query;
+mod records_read;
+mod records_subscribe;
+mod records_write;
