--- src/messages.rs
@@ -1,7 +1,0 @@
-//! # Messages Handlers
-//!
-//! Messages handlers handle incoming `Messages` interface messages.
-
-mod query;
-mod read;
-pub mod subscribe;
