--- src/records.rs
@@ -1,11 +1,0 @@
-//! # Records Handlers
-//!
-//! Records handlers handle incoming `Records` interface messages.
-
-mod delete;
-mod integrity;
-mod protocol;
-mod query;
-mod read;
-mod subscribe;
-pub mod write;
