+++ tests/records_write.rs
@@ -1,5 +1,7 @@
 //! Records Write
 
+#![cfg(all(feature = "client", feature = "server"))]
+
 use std::io::Cursor;
 use std::sync::LazyLock;
